+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - Custom About window with version info and links (Website, GitHub, Documentation)
+- Build script and C bridge module for libmongoc (MongoDB C Driver) static libraries
 
 ## [0.9.2] - 2026-02-28